done all

+++ client/src/app/books/books.component.ts
@@ -4,6 +4,7 @@ import { Observable } from 'rxjs';
 import { book } from '../types/book';
 import { AsyncPipe, CommonModule } from '@angular/common';
 import { RouterLink } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-books',
@@ -17,18 +18,25 @@ export class BooksComponent implements OnInit{
   bookService = inject(BooksService);
 index: any;
 
+  constructor(private toasterService: ToastrService) { }
+
   ngOnInit(): void {
     this.books$ = this.bookService.getBooks()
     
   }
 
   deleteBook(id: number) {
-    this.bookService.deleteBook(id).subscribe({
-      next: () => {
-        // Refresh the list after deletion
-        this.books$ = this.bookService.getBooks();
-      },
-      error: (err) => console.log(err)
-    });
+    if (confirm('Are you sure you want to delete this book?')) {
+      this.bookService.deleteBook(id).subscribe({
+        next: () => {
+          this.books$ = this.bookService.getBooks();
+          this.toasterService.success('Book deleted successfully!');
+        },
+        error: (err) => {
+          this.toasterService.error('Failed to delete book!');
+          console.log(err);
+        }
+      });
+    }
   }
 }
